+++ CHANGELOG.md
@@ -56,11 +56,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Bump 4DO to support threaded DSP emulation & lightgun
 - Activate 4DO package on Rpi3
 - Bump Amiberry to 2.24 (SDL2). Compatible Odroid XU4.
-- Add Amige CD32 system
+- Add Amiga CD32 system
 - Add AROS bios as default Amiga bios (2019 version)
 - Add configuration for static ip addresses on wifi interfaces
 - Bump ScummVM/ResidualVM (January 2019)
 - Bump almost all libretro cores
+- Package names updated (catsfc -> snes9x2005, pocketsnes -> snes9x2002, snes9x_next -> snes9x2010, pce -> mednafen_pce_fast, vb -> mednafen_vb, imame -> mame2000, mame2003 as core name instead of mame078, fbalpha instead of fba)
 
 ## [18.07.13] - 2018-07-13
 - Bump desmume to desmume2015